refactor(admin): extract the Authentication panel from ConfigTab
Second of the three independently-saving panels (plan 013, phase 1). Moves
the Authentication + OAuth/SAML Providers panel — handleSaveAuthMethods,
handleAddProvider, handleUpdateProvider, handleDeleteProvider,
handleEditProvider, handleImportSamlMetadata, providerValidationError and
their state — into components/admin/config/AuthPanel.tsx.

This is the panel that can lock every user out of the instance, so the two
guards are preserved verbatim and now covered by tests in isolation:

  - the auth-methods floor (plan 003): the last enabled method's checkbox
    stays disabled, including after the other method is unchecked
  - the "***" client-secret sentinel: Edit still prefills "***" and Save
    still echoes it, so the stored secret is kept rather than blanked
  - providers stay addressed by stable id (plan 011)
  - the panel keeps id="cfg-auth" (the setup-checklist jump target)

The panel JSX is byte-identical to what it replaced apart from two reads of
`cfg.oauth_providers` becoming the `providers` prop (verified by diff).

Five props, no setters passed down. authMethods is seeded from the loaded
config as an initial value, not a controlled one, because provider writes
re-read the whole config and must not reset an in-progress selection — which
is exactly what the old shared-state version did.

ConfigTab.tsx: 1,957 → 1,518 lines. handleSaveConfig and the sticky-save
cluster remain untouched. All 11 ConfigTab characterization tests still pass
unchanged, plus 13 new AuthPanel tests. 48 files / 370 tests; ConfigTab chunk
87,780 → 87,802 bytes.
